Refactor sphinx candidate list rendering
This change now uses the new schema to render candidate
list including ircname.

Moreover this change introduces an exceptions.txt file to
manage candidacies exception where a candidate isn't the
git commit author of a candidacy statement. In such case
the candidacy needs to be manually verified and the candidate
fullname needs to be added to the exceptions.txt.

Change-Id: I2e59f4880d34cdb09321dbc110ecf5a103f81dff