Allow human readable yaml
The default yamllint rules allow for only a single space after a key.  It's
kinda nice to be able to space align the dates to make them slightly more
readable by a human.  Add a yamllint file that sets that up, at the same time
demote these from error to warning.

Change-Id: Iab4c2a466146f4c440e6552a1341fc4474c62502