Allow create-directories to be run multiple times
Currently the create-directories tool can only be run once as if the
candidates/<election>/ directory exists it will abort.

If a project is approved by the TC after the directories are created but
before the election start that candidates directory needs to be created
by hand.  Instead change create-directories to only create the
files/directories as needed so it can be run several times.

Change-Id: I62bba67447e2ed91138d6e099b45411cba72be36