Use the tag when grabbing the projects list
The helper that added the tag to the cgit url was incorrectly appending
a simple "?tag" rather than "?h=tag"  cgit treats the tag as an unknown
option and is silently ignoring it.

Correct that.

The impact on the election is that any candidacies may have validated
may have included extra-atcs or repos added by the TV after the tagging
of the repo.  Checking the governance repo we see:

$ git log --no-merges --oneline aug-2017-elections..origin/master -- reference/projects.yaml
9b64bcb Adds oswin-tempest-plugin for Winstackers

This change clearly only affects the Winstackers candidates.

Change-Id: I1dfc47cb41aac4ce70d280e860ee4536cb90fd17