Correct start date for Pike+Queens contributions
The Pike release cycle started on or around the date of the Ocata
release, not the Pike release. Since the TC charter defines the PTL
electorate (APCs) as "Individual Members who committed a change to a
repository of a project over the last two 6-month release cycles" we
need to include back as far as February 22, 2017.

In prior elections, the start was considered the end of the release
day (23 hours 59 minutes UTC), so preserve that quirk for
consistency. We can revisit that choice in later elections.

https://governance.openstack.org/tc/reference/charter.html#voters-for-ptl-seats-apc
https://releases.openstack.org/ocata/index.html

Change-Id: I08fb6f994d3d05d26f5b0fd7495056e3065f38e1