Handle failed lookups from the OSF member DB
If a lookup for a member find no matches (perhaps an incorrect email
address) the resulting JSON has an empty 'data' element.  The existing
code assumed it would be missing altogether.

So check for an item in the list before trying to index the data.

While we're doing this, use the same code style for both get_fullname()
and get_irc() to remove points of confusion.

Change-Id: I4c899df5f93c27f5e5390c93e88741d24c0282d8