Prepare Stein PTL election
Setting PTL Election. Release is at: 2018-08-30
Latest possible completion is at: 2018-08-09
Moving back to Tuesday: 2018-08-07
PTL Nominations from 2018-07-24T23:45 to 2018-07-31T23:45
PTL Election from 2018-07-31T23:45 to 2018-08-07T23:45

The date for the beginning of the queens development cycle was
approximately the beginning of RC1 week which is about when we branch.
The end of Rocky has been set to match the email deadline, which gives
the officials the week of nominations to prepare and validate rolls.

I have also removed myself as an election official for the typical
reasons.

Change-Id: I4a775c95064ea81106b906e6f0a2e0ec915ae714