Handle projects that aren't resolved at conclusion of the election
This cycle we have a few projects that are leaderless and the haven't
been completely closed off during the election window.

Add the ability for election officials to provide clarifying text in
this case.  As each project is likely to be slight different we have
opted against a one size fits all statement.

Change-Id: I5d82c048e1d38a3d73d8390a98ee9f948e252522