Factor out the logic to collect the candidate files
Currently we use glob.glob() to get a list of files to validate.  We
make a simple choice to only use 'TC' for TC elections and '*' for PTL
elections.  This mostly works when the TC election is after the PTL
election.  Consider the following tree:
   candidates/some_election/TC/candidate@one
   candidates/some_election/TC/candidate@two
   candidates/some_election/project_one/candidate@three
   candidates/some_election/project_one/candidate@four
   candidates/some_election/project_two/candidate@five

The current logic will check only candidate@one and candidate@two for a
TC election but all files for a PTL election if the PTL election happens
after the TC election.

So factor out the file system logic so we can correctly include/exclude
the TC based on the election type.

Change-Id: I493d3663afda9d5e849176b3e6a9d042e0af7f3e