Don't validate .placeholder
When creating a new election we need to add a directory per project
(including the 'TC' as a pseudo project) we can't just create an empty
directory in git so we add a '.placeholder' file.

We have recently added a voting job that will run try to validate all
candidate files added in a review.  This includes the placeholder file.
CLearly this file will not pass validation and cause the job to fail.

Adding .placeholder to the zuul irrelevant-files variable wont work, so
instead filter .placeholder from the interesting files.

Change-Id: Ibd2224a5479d4ab9afbac227bac38080359a98bf