Reduce the differences between TC and PTL elections
It turns out that structure of the TC and PTL elections is very similar
and therefore the template choices are basically the same.  Remove the
overly complex sub-parser and always use the election_type from the YAML
configuration.

Also many of the fmt_args where shared, or harmless to share between
each election type.  pull the similar items out and only adds the items
specific to each election type

Change-Id: Ifcc8a376e7d8d9a79ae7a9191cb2452f8cfbae43