Include election process in rendered documentation
Create a stub process document which includes the readme so it will
be incorporated into the published election site content. Link to it
from within the main index as well. Overhaul the semi-compliant
reStructuredText syntax in README.rst so it will parse correctly,
and add some additional markup for readability. Also include a note
in the main page section on officials with a call for volunteers.

Change-Id: If78b23a5c8d385d58a2c3125a704d15b5f04a0f1