Officially call it the "U" election
Because it will still take some time for the OSF to perform
legal/trademark reviews on the results of the U cycle naming poll
and we want to be able to move ahead with accepting nominations for
upcoming elections, switch to just calling this cycle by the
required initial letter we already know. In official communications
we can capitalize it as "U" but for the purposes of configuring our
release automation the previous names were lower-case so use "u"
here instead.

Change-Id: I0310c0cdbc6492f375e8a1364873ee80edda8e2b