Truncate Ussuri election schedule
Since the combined PTL/TC elections resulted in no contested seats,
all candidates were acclaimed and no polls were required. Because of
this, there is no need for a campaigning week or polling week in the
schedule. Shorten these phases to one minute each so that the site
renders the expected state.

Change-Id: Iff1311a8fbcedf3cd5d39e381d68f10bad376d0e