Update tag for September 2019 elections
This is just a bookkeeping change to reflect that the TC is
switching to using OpenStack Release Management automation for
governance repository tags, and these need to be typical numerical
tags rather than the election-specific tags we've used in the past.
The value of this configuration setting is only presently used when
generating electoral rolls, which we did not need to do for the last
election due to there being no polls required. As a result, it is
safe to update this after the fact to leave a record correlating to
the relevant point in the governance repository which would have
been used if we did need to generate rolls.

Depends-On: I619ca1e8c2710ae020b4be4683f96ed345539d40
Change-Id: Ib06de473166761c12ea4955802d50b497287614a