Include contributors to repos for SIGs in TC rolls
It's been the intent ever since the TC started keeping a record of
what Git repositories are controlled by official SIGs, that the
contributors to those repositories should be eligible to vote in TC
elections. Add that feature, along with a command-line option for
specifying a non-default file in the event the official one needs to
be overridden, and use the same zeroing logic as is used for
skipping the legacy file for retired repositories, for the same
reasons.

Change-Id: I65d2a620a0e4b81d0ef685a7fedb0e7800cbff9f