Include aggregate change revisions in owner data
In order to make some calculations around revision-per-change rates,
add a revision count for merged changes alongside the existing
merged change count. This is data we already get from the Gerrit
API, so it shouldn't impact performance, and it doesn't alter any
existing fields in the output (only adds a new one).

Change-Id: I2e3129efc5cd96e5c93815b6d0e699f04a838bef