Set Victoria  configuration
Based on the email thread back in December, we settled roughly
on a set of dates, but concluded we should bump them back
an extra week so that PTL elections can be concluded before
we open the Victoria master branch. The necessary
TC charter changes to update term lengths have also been
made.

Change-Id: I66872e55b31319a00cfee795898f2990dfc87fe7