Add extra-ATCs from sigs-repos into the electorate
Without this patch, extra-atcs that are defined in a SIG will not be
taken into consideration for electorate rolls.

This is a problem, as we intend to move i18n as a SIG, so we would
lose their atc as part of the electorate.

This fixes it by editing the spaghetti code for SIGs. It was already
simulating the structure of a project named "SIG". So it made sense
to extend it with 'extra-atcs'.

Therefore to initialise extra-atcs for SIGs, we just have to
have something like this in sigs-repos:

```
signame:
  - repo: repo_location
    extra-atcs:
    - name: extra_atc_fullname
      email: extra_atc_email
      expires-in: extra_atc_expiry
      comment: extra_atc_comment
```

Yes, it means that a SIG need to own a repo for this.
This is not a problem at the moment, AFAIK.
If this becomes a problem, we could think of removing the sigs-repos file,
and just make SIGs some kind of project in governance, simplifying the code.

Depends-On: https://review.opendev.org/721605
Change-Id: I5e1c42c64f5e76dafb9420aa1f3363890b1b9d65