Add a combined_end_nominations E-mail template
Since the last time we ran a combined election there were no runoffs
needed, nobody had yet created an end_nominations template for the
combined model. This one is a mashup of the existing PTL and TC
templates, with some patterns borrowed from combined templates for
other events. Also included is a typo fix for the
combined_nominations_last_days template which I just noticed when
referencing it.

Change-Id: Icfbff4ec284f227d2c575f43cf97e27190bca62e