Set up TC Vacancy Special Election
During the TC meeting today[*], an immediate special election was
authorized for filling the seat which was left vacant when the prior
election did not receive sufficient nominations to fill all
available seats. We will be holding an election for one remaining
seat in the 2021-Q1 to 2022-Q1 term. As discussed in the meeting
there will be two weeks for nominations (during which time
candidates can also campaign if desired), followed by one week of
voting if we have more than one candidate.

The election setup reflects the full 5 seats for the term, even
though four are already filled, and will display the four already
elected candidates for the term so that we can reuse the same
nominations directory. When the nominations period concludes, only
the candidates not already elected for the term will be added to a
new poll for a single seat.

[*] http://eavesdrop.openstack.org/meetings/tc/2021/tc.2021-03-25-15.00.log.html#l-118

Change-Id: I2f78e3fa3474c1473b7805dc574a69e150e0bc6e