Add election_results to cycle data
The Jinja templates for formatting the election results expect an
'election_results' property to be loaded from the results yaml files to
generate election results.

There have been a few election cycles for PTLs now that have not needed
a election due to only having one candidate. In these cases, the yaml
files for the ptl results of those election cycles did not include the
empty 'election_results' property. This adds it to the missing data sets
to allow the docs job to pass.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>
Change-Id: I759b06a6b977299873ec8ab1f7e727d8b7071518