Fix setup-election-config script for TC election dates
setup-election-config script is still using the old
dates for TC election
"The election is held no later than 6 weeks prior to each
 OpenStack Summit (on or before 'S-6' week)"

and set the offset of 6 weeks of summit or release. As we do not
have summit in every cycle so we need to update it to take
reference from release dates.

Also TC charter is changed 2 year back about new recommended dates
for TC election, which is:

"The election is held no earlier than 6 weeks and no later
than 4 weeks prior to each cycle final release date (between 'R-6' and 'R-4'
week)"

- https://review.opendev.org/c/openstack/governance/+/699277/6/reference/charter.rst

This commit fix the script with the offset as 4 weeks from release date
so that generated dates will be as per the TC charter.

Change-Id: Id43b67add3ec405bd94c64a9d2d324247914031a