Add 'combined' election support in setup-election-config
setup-election-config script is used to generate the
election dates. As we started doing the combined election
of PTL and TC, we need to add that type of election support
in this tool which will help to generate dates for any
type of election.

Below the is generated data will looks like:

release: yoga
election_type: combined
tag: oct-2021-elections
tc_seats: 4
timeframe:
  name: Wallaby-Xena
  start: 2020-09-25T00:00
  end: 2021-08-24T00:00
  email_deadline: 2021-08-24T00:00
timeline:
- name: TC & PTL Nominations
  start: 2021-08-17T23:45
  end: 2021-08-24T23:45
- name: TC Campaigning
  start: 2021-08-24T23:45
  end: 2021-08-31T23:45
- name: TC & PTL Election
  start: 2021-08-31T23:45
  end: 2021-09-07T23:45

Change-Id: Ib97f7ba794078ed78c9241dce7d40b3fc3519948