Automate the 'date' in setup_election_config
Current we need to pass 'date' as mandatory argument
in setup_election_config CLI and it is sometime confusion
on what date to pass.

This commit automatically select the end date of release if
date is not passed. This will help to generate the correct
election dates without worry about release date to pass.

Change-Id: I444885800b26ca84689d550f81bbe349eae3bd30