Remove hardcoded extlinks link from doc/source/conf
Doc job is failing because sphinx 4.4.0
raise error for hardcoded extlinks link
- https://github.com/sphinx-doc/sphinx/pull/9800

Warning, treated as error for extlinks defined in the
doc/source/conf.py

To fix those error we need to replace all the 'https:/opendev.org/openstack/'
which is defined as extlinks in all the ptl.rst files from previous
release elections too. I think we just remove to define it as extlinks
and keep full complete url in doc which does not harm.

Change-Id: I263da27b3c39a5816e6c8961027337b02e7331d7