Migrate election tooling to release versions
From Antelope (2023.1) onwards, primary identifier
of any release is release version instead of name[1],
we should use release versions in the election tooling
also.

This migrate the election tooling to use the release version
As release repo is not fully migrated to release version yet,
we still need to use the release name to fetch the data from
release repo.

Keeping relaese version as float so that we edit the
configuration file with release version a float not
with string even string is passed.

All the election tools/script/flow has been tested with it and
it is working now.

[1] https://governance.openstack.org/tc/reference/release-naming.html

Change-Id: Ia8a60a9514b1dca5ac97ed8099c34c6c6c7d24e2