Update package metadata and tox4 setup
The current tox setup results in an 'editable-wheel' install which,
while it works, generates unhelpful warnings.  Switch to the 'editable'
type which is what you got under tox <= 4.

While we are reducing noise in the logs, correctly setup
openstack_elections.templates as a module.  We need it as a module as
we rely on Jinja2 to be able to use the module loader to locate the
templates.

Change-Id: Icbdca2f72c0777c587d6f95b3fab1911be179f97