Fix the tag to be used for setting election config
When we start the election configuration for dates and
other deadlines, we use the dummy tag generated with election
month and which gets updated once governance release the
actual tag of semver. It is confusing that initial tag
generated by setup_election_config cmd is actual tag or just
a placeholder.

Governance repo has switched to semver type tag since 2018
- https://github.com/openstack/governance/tags?after=oct-2017-elections

Election repo started the below process since 2019:
1. Add placeholder/dummy tag in election config (election month tag generated by script)
- https://review.opendev.org/c/openstack/election/+/661673
2. On email_deadline, release a new tag of governance repo
- https://review.opendev.org/c/openstack/releases/+/680691
3. Update the released tag in election config
- https://review.opendev.org/c/openstack/election/+/681274

This commit does not change the existing process and make the placeholder tag more
clear so that we know that we need to release a new tag in governance at the email_deadline
and update here.

Change-Id: Id8d5e4689ad2374a197e63e7fb41c6b5c265ff5c