Update job config.
In Switch to 2023.2 testing runtime py version (7f56f206e6) we added
the openstack-python3-jobs template to the zuul config.  This doesn't
work for the elections repo as we don't have a pep8 target, and we also
do not want to add publish-openstack-python-branch-tarball to the
promote pipeline.

Instead just opencode the python jobs.

Change-Id: I289f16d95e19da2e546300dd2a2376f2ccd3ac86