Convert ATC to AC term
TC passed a resolution in 2021[1] to convert the term
ATC to AC and this is not just wording change but also
it change the definition of it which include the AUC also
in AC. Previously ATC and AUC were two different things.

Reflecting the TC resolution in election tooling also.

[1] https://governance.openstack.org/tc/resolutions/20210602-atc-renamed-to-ac.html

Depends-On: https://review.opendev.org/c/openstack/governance/+/893831
Change-Id: Ic0570ae9a2bb2b31538179d3246b4475cdbeac45