[templates] Update links
Since switching to the release_id instead of the release_name we're no
longer generating HTML anchors as expected.

The generated RST:
'
2024.1 TC Candidates
====================
'

Doesn't generate #2024.1-tc-candidates, instead it generates #tc-candidates.
This is because docutils 'ids' cannot start with numerics ([1] is the
implementation).  So we need to change the text/anchors.  It seems like
including release specific names or ids isn't actually helpful.  Let's
drop them.

[1] https://repo.or.cz/docutils.git/blob/refs/heads/master:/docutils/docutils/nodes.py#l2226

Change-Id: I4c48ade272ae3d9bd4a97fba7d30f909aac696f9