Switch foundation membership check to new style
With the dissolution of the OpenInfra Foundation incorporated in
Delaware and move of operations to a directed fund within the Linux
Foundation, former Individual Members were asked to reassert
membership within the new organization. In order to track both
within the backing database, the old foundation-members group is no
longer relevant and instead a membership_type field has been added
with a value of "Individual" indicating someone is an Individual
Project Member for the new directed fund. Because of this, the API
query we use to identify membership status in the foundation is
updated to reflect the change.

Change-Id: I18842e99aecb71cb03668e321eef7e49e6b2ec47
Signed-off-by: Jeremy Stanley <fungi@yuggoth.org>