Use mergedafter and mergedbefore Gerrit filters
A simple change to speed up Gerrit queries by reducing the number of
changes returned. Now we no longer need to get all changes with
updates since the after date and check whether and when they merged,
since Gerrit has added support for listing just changes merged after
and before specific times.

There is likely a lot of additional unnecessary checking that can be
cleaned up after this, to reduce the amount of work Python is doing,
but limiting the query responses will make the largest dent in total
runtime so we start there.

Change-Id: I6aa5f1a3808b2241699fa6924c07e17e955d0342
Signed-off-by: Jeremy Stanley <fungi@yuggoth.org>